pyproject.toml
[project]
name = "console_ppt"
version = "0.1.0"
description = "A terminal-based PPT presentation tool from Markdown"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.47.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
console_ppt = "console_ppt.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["console_ppt"]
